✨ feat: 인기 탭 컴포넌트 분리 (#35)
